SQL File Uploads
Tried to create a script but it was being difficult to connect to oracle so I went in converted the SkiResort.csv files to a .txt file, then I created the tables and uploaded them to Oracle. They were granted public access and you should be able to view them using pruiz2.[TableName], but use Member, SkiPass, etc.